Refactor build compilation in anticipation of scripting support.
